move BuildHTTPClient from shared httputil to catalogmetadata/client
